[2020-09-23] Event API_26 : Refactoring Test Code
 - 성공/실패 Test Case 분리 : *Fail* / *Success*
 - Event 객체 생성 후 조회 부 분리 : EventDomainGenerator.java
 - Auth 정보 조회 부 분리 : AuthInfoGenerator.java
