+++ src/test/java/io/api/event/common/BaseControllerTest.java
@@ -1,6 +1,8 @@
 package io.api.event.common;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import io.api.event.common.auth.AuthInfoGenerator;
+import io.api.event.common.event.EventDomainGenerator;
 import io.api.event.config.ApplicationProperties;
 import io.api.event.config.RestDocsConfiguration;
 import io.api.event.config.TestConstants;
@@ -16,6 +18,8 @@ import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
 
+import javax.annotation.Resource;
+
 /**
  * 2020-09-15
  *  - Controller Test Code에서 반복적으로 사용되는 설정 부를 추출 후 해당 클래스를 상속 받아 코드 중복 제거
@@ -24,10 +28,11 @@ import org.springframework.test.web.servlet.MockMvc;
 @SpringBootTest // 통합테스트 : Application 설정을 이용하여 Test 환경 구성
 @AutoConfigureMockMvc // @SpringBootTest annotation을 이용한 통합테스트 진행 시 해당 TC내에서 MockMvc를 주입하기위한 annotation
 @AutoConfigureRestDocs
-@Import(RestDocsConfiguration.class)
+@Import({RestDocsConfiguration.class, AuthInfoGenerator.class, EventDomainGenerator.class})
 @ActiveProfiles(TestConstants.TEST) // Test Application 환경 설정
 @Disabled // jUnit4의 @Ignore 대체 annotation : Test를 가지고 있는 class로 간주되지 않도록 설정
 @TestMethodOrder(MethodOrderer.Alphanumeric.class) // Test method name을 이용한 실행순서 정렬
+//@Import(AuthInfoGenerator.class)
 public class BaseControllerTest {
 
     @Autowired
@@ -41,5 +46,11 @@ public class BaseControllerTest {
 
     @Autowired
     protected ApplicationProperties applicationProperties;
+
+    @Resource
+    protected AuthInfoGenerator authInfoGenerator;
+
+    @Resource
+    protected EventDomainGenerator eventDomainGenerator;
 
 }
