+++ src/test/java/io/api/event/controller/EventControllerSuccessTest.java
@@ -2,8 +2,6 @@ package io.api.event.controller;
 
 import io.api.event.common.BaseControllerTest;
 import io.api.event.domain.dto.event.EventDto;
-import io.api.event.domain.entity.account.Account;
-import io.api.event.domain.entity.account.AccountRole;
 import io.api.event.domain.entity.event.Event;
 import io.api.event.domain.entity.event.EventStatus;
 import io.api.event.repository.EventRepository;
@@ -18,14 +16,13 @@ import org.springframework.hateoas.MediaTypes;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders;
-import org.springframework.security.oauth2.common.util.Jackson2JsonParser;
 import org.springframework.test.web.servlet.ResultActions;
 
 import java.nio.charset.StandardCharsets;
 import java.time.LocalDateTime;
-import java.util.Set;
 import java.util.stream.IntStream;
 
+import static io.api.event.common.document.DocumentFormatGenerator.getDateTimeFormat;
 import static org.springframework.restdocs.headers.HeaderDocumentation.*;
 import static org.springframework.restdocs.hypermedia.HypermediaDocumentation.linkWithRel;
 import static org.springframework.restdocs.hypermedia.HypermediaDocumentation.links;
@@ -33,7 +30,6 @@ import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.docu
 import static org.springframework.restdocs.payload.PayloadDocumentation.*;
 import static org.springframework.restdocs.request.RequestDocumentation.parameterWithName;
 import static org.springframework.restdocs.request.RequestDocumentation.pathParameters;
-import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.httpBasic;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
@@ -43,7 +39,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
  * https://woowabros.github.io/experience/2018/12/28/spring-rest-docs.html
  */
 //@AutoConfigureRestDocs()
-public class EventControllerTest extends BaseControllerTest {
+public class EventControllerSuccessTest extends BaseControllerTest {
 
     @Autowired
     EventRepository eventRepository;
@@ -59,7 +55,6 @@ public class EventControllerTest extends BaseControllerTest {
         eventRepository.deleteAll();
         accountRepository.deleteAll();
     }
-
 
     @Test
     @TestDescription("Spring HATEOAS, Spring REST DOCS를 이용한 API 응답, 전이 가능한 Link정보, Docs 생성 유무 확인")
@@ -82,7 +77,7 @@ public class EventControllerTest extends BaseControllerTest {
         // When
         String urlTemplate = "/api/events";
         ResultActions resultActions = mockMvc.perform(post(urlTemplate)
-                .header(HttpHeaders.AUTHORIZATION, getBearerToken())
+                .header(HttpHeaders.AUTHORIZATION, authInfoGenerator.getBearerToken())
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .accept(MediaTypes.HAL_JSON)
                 .characterEncoding(StandardCharsets.UTF_8.name())
@@ -125,10 +120,10 @@ public class EventControllerTest extends BaseControllerTest {
                         requestFields(
                                 fieldWithPath("name").description("Name of new event"),
                                 fieldWithPath("description").description("Description of new event"),
-                                fieldWithPath("beginEnrollmentDateTime").description("Date time of begin enrollment of new event"),
-                                fieldWithPath("closeEnrollmentDateTime").description("Date time of close enrollment of new event"),
-                                fieldWithPath("beginEventDateTime").description("Date time of begin of new event"),
-                                fieldWithPath("endEventDateTime").description("Date time of close of new event"),
+                                fieldWithPath("beginEnrollmentDateTime").description("Date time of begin enrollment of new event").attributes(getDateTimeFormat()),
+                                fieldWithPath("closeEnrollmentDateTime").description("Date time of close enrollment of new event").attributes(getDateTimeFormat()),
+                                fieldWithPath("beginEventDateTime").description("Date time of begin of new event").attributes(getDateTimeFormat()),
+                                fieldWithPath("endEventDateTime").description("Date time of close of new event").attributes(getDateTimeFormat()),
                                 fieldWithPath("location").description("Location of new event").optional(),
                                 fieldWithPath("basePrice").description("Base Price of new event"),
                                 fieldWithPath("maxPrice").description("MaxPrice of new event"),
@@ -164,89 +159,11 @@ public class EventControllerTest extends BaseControllerTest {
     }
 
     @Test
-    @TestDescription("JSR303 Annotation을 이용한 입력값이 없는 요청의 400 Bad Request 처리")
-    @DisplayName("Create Event API : 입력값이 없는 요청")
-    public void createEventAPI_EmptyRequest_Test() throws Exception {
-        // Given
-        EventDto eventDto = new EventDto();
-
-        // When
-        String urlTemplate = "/api/events";
-        ResultActions resultActions = mockMvc.perform(post(urlTemplate)
-                .header(HttpHeaders.AUTHORIZATION, getBearerToken())
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .accept(MediaTypes.HAL_JSON_VALUE)
-                .characterEncoding(StandardCharsets.UTF_8.name())
-                .content(objectMapper.writeValueAsString(eventDto))
-        );
-
-        /** Then : Check list
-         * - 응답 코드 확인 : 400 Bad Request
-         * - 응답 항목 내 오류 정보 확인 : field/global Error (objectName, defaultMessage, code)
-         * - 응답 항목 내 index link 항목 확인
-         */
-        resultActions
-                .andDo(print())
-                .andExpect(status().isBadRequest())
-                .andExpect(jsonPath("content").isArray())
-                .andExpect(jsonPath("content[0].objectName").exists())
-                .andExpect(jsonPath("content[0].field").exists())
-                .andExpect(jsonPath("content[0].code").exists())
-                .andExpect(jsonPath("content[0].defaultMessage").exists())
-                .andExpect(jsonPath("_links.index").exists())
-        ;
-    }
-
-    @Test
-    @TestDescription("CustomValidator를 이용한 입력값이 유효하지 못한 요청의 400 Bad Request 처리")
-    @DisplayName("Create Event API : 입력값이 유효하지 못한 요청")
-    public void createEventAPI_WrongParameterRequest_Test() throws Exception {
-        // Given
-        Event event = Event.builder()
-                .name("루나소프트 생활 체육회")
-                .description("제 2회 루나 배 풋살 대회")
-                .beginEnrollmentDateTime(LocalDateTime.of(2020, 8, 6, 9, 30 ))
-                .closeEnrollmentDateTime(LocalDateTime.of(2020, 8, 7, 9, 30 ))
-                .beginEventDateTime(LocalDateTime.of(2020, 8, 6, 19, 0))
-                .endEventDateTime(LocalDateTime.of(2020, 8, 13, 22, 0))
-                .basePrice(300)
-                .maxPrice(200)
-                .limitOfEnrollment(0)
-                .location("서울시 강남구 일원동 마루공원 풋살장 1면")
-                .build();
-
-        // When
-        String urlTemplate = "/api/events";
-        ResultActions resultActions = mockMvc.perform(post(urlTemplate)
-                .header(HttpHeaders.AUTHORIZATION, getBearerToken())
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .accept(MediaTypes.HAL_JSON)
-                .characterEncoding(StandardCharsets.UTF_8.name())
-                .content(objectMapper.writeValueAsString(event))
-        );
-
-        /** Then : Check list
-         * - 응답 코드 확인 : 400 Bad Request
-         * - 응답 항목 내 오류 정보 확인 : field/global Error (objectName, defaultMessage, code)
-         * - 응답 항목 내 index link 정보 확인
-         */
-        resultActions.andDo(print())
-                .andExpect(status().isBadRequest())
-                .andExpect(jsonPath("content[0].objectName").exists())
-                .andExpect(jsonPath("content[0].defaultMessage").exists())
-                .andExpect(jsonPath("content[0].code").exists())
-                .andExpect(jsonPath("content[0].defaultMessage").exists())
-                .andExpect(jsonPath("_links.index").exists())
-        ;
-    }
-
-
-    @Test
     @TestDescription("Spring HATEOAS, Spring REST DOCS를 이용한 API 응답, 전이 가능한 Link정보, Docs 생성 유무 확인")
     @DisplayName("Get Event List API : 이벤트 목록 조회 요청")
     public void getEventListAPI_Test() throws Exception {
         // Given
-        IntStream.range(0, 30).forEach(this::generatedEvent);
+        IntStream.range(0, 30).forEach(eventDomainGenerator::generatedEvent);
 
         // When
         String urlTemplate = "/api/events";
@@ -297,11 +214,11 @@ public class EventControllerTest extends BaseControllerTest {
     @DisplayName("Get Event API : 이벤트 조회 요청")
     public void getEventAPI_Test() throws Exception {
         // Given
-        Event event = this.generatedEvent(100);
+        Event event = eventDomainGenerator.generatedEvent(100);
 
         // When :
         String urlTemplate = "/api/events/{id}";
-        ResultActions resultActions = mockMvc.perform(RestDocumentationRequestBuilders.get("/api/events/{id}", event.getId())
+        ResultActions resultActions = mockMvc.perform(RestDocumentationRequestBuilders.get(urlTemplate, event.getId())
                 .accept(MediaTypes.HAL_JSON_VALUE)
                 .contentType(MediaType.APPLICATION_JSON)
                 .characterEncoding(StandardCharsets.UTF_8.name())
@@ -361,36 +278,11 @@ public class EventControllerTest extends BaseControllerTest {
     }
 
     @Test
-    @TestDescription("요청 정보에 해당하는 Event 객체 조회 실패 시, 404 Not Found 처리")
-    @DisplayName("Get Event API : 존재 하지 않는 이벤트 조회 요청")
-    public void getEventAPI_NotFount_Test() throws Exception {
-        // Given
-        String urlTemplate = "/api/events/12134";
-
-        // When
-        ResultActions resultActions = mockMvc.perform(get(urlTemplate)
-                .accept(MediaTypes.HAL_JSON_VALUE)
-                .contentType(MediaType.APPLICATION_JSON)
-                .characterEncoding(StandardCharsets.UTF_8.name())
-        );
-
-        /** Then : Check list
-         * - 응답 코드 확인 : 404 Not Found
-         * - 응답 항목 내 index link 항목 확인
-         */
-        resultActions
-                .andDo(print())
-                .andExpect(status().isNotFound())
-                .andExpect(header().exists(HttpHeaders.LOCATION))
-        ;
-    }
-
-    @Test
     @TestDescription("Spring HATEOAS, Spring REST DOCS를 이용한 API 응답, 전이 가능한 Link정보, Docs 생성 유무 확인")
     @DisplayName("Update Event API : 이벤트 수정 요청")
     public void updateEventAPI_Test() throws Exception {
         // Given
-        Event event =  this.generatedEvent(100);
+        Event event =  eventDomainGenerator.generatedEvent(100);
         EventDto eventDto = this.modelMapper.map(event, EventDto.class);
         String updatedEventName = "updated Event Name";
         eventDto.setName(updatedEventName);
@@ -398,7 +290,7 @@ public class EventControllerTest extends BaseControllerTest {
         // When
         String urlTemplate = "/api/events/{id}";
         ResultActions resultActions = mockMvc.perform(put(urlTemplate, event.getId())
-                .header(HttpHeaders.AUTHORIZATION, getBearerToken())
+                .header(HttpHeaders.AUTHORIZATION, authInfoGenerator.getBearerToken())
                 .accept(MediaTypes.HAL_JSON_VALUE)
                 .contentType(MediaType.APPLICATION_JSON)
                 .characterEncoding(StandardCharsets.UTF_8.name())
@@ -424,150 +316,4 @@ public class EventControllerTest extends BaseControllerTest {
         ;
     }
 
-    @Test
-    @TestDescription("JSR303 Annotation을 이용한 입력값이 없는 요청의 400 Bad Request 처리")
-    @DisplayName("Update Event API : 입력값이 없는 요청")
-    public void updateEventAPI_EmptyRequest_Test() throws Exception {
-        // Given
-        Event event = this.generatedEvent(100);
-        EventDto eventDto = new EventDto();
-
-        // When
-        String urlTemplate = "/api/events/{id}";
-        ResultActions resultActions = mockMvc.perform(put(urlTemplate, event.getId())
-                .header(HttpHeaders.AUTHORIZATION, getBearerToken())
-                .accept(MediaTypes.HAL_JSON_VALUE)
-                .contentType(MediaType.APPLICATION_JSON)
-                .characterEncoding(StandardCharsets.UTF_8.name())
-                .content(this.objectMapper.writeValueAsString(eventDto))
-        );
-
-        // Then
-        resultActions.andDo(print())
-                .andExpect(status().isBadRequest())
-                .andExpect(jsonPath("content").isArray())
-                .andExpect(jsonPath("content[0].objectName").exists())
-                .andExpect(jsonPath("content[0].field").exists())
-                .andExpect(jsonPath("content[0].code").exists())
-                .andExpect(jsonPath("content[0].defaultMessage").exists())
-                .andExpect(jsonPath("_links.index").exists())
-        ;
-    }
-    
-    @Test
-    @TestDescription("CustomValidator를 이용한 입력값이 유효하지 못한 요청의 400 Bad Request 처리")
-    @DisplayName("Update Event API : 입력값이 유효하지 못한 요청")
-    public void updateEventAPI_WrongParameterRequest_Test() throws Exception {
-        // Given
-        Event event =  this.generatedEvent(100);
-        EventDto eventDto = this.modelMapper.map(event, EventDto.class);
-        eventDto.setBasePrice(20000);
-        eventDto.setMaxPrice(10000);
-        String updatedEventName = "updated Event Name";
-        eventDto.setName(updatedEventName);
-
-        // When
-        String urlTemplate = "/api/events/{id}";
-        ResultActions resultActions = mockMvc.perform(put(urlTemplate, event.getId())
-                .header(HttpHeaders.AUTHORIZATION, getBearerToken())
-                .accept(MediaTypes.HAL_JSON_VALUE)
-                .contentType(MediaType.APPLICATION_JSON)
-                .characterEncoding(StandardCharsets.UTF_8.name())
-                .content(this.objectMapper.writeValueAsString(eventDto))
-        );
-
-        // Then
-        resultActions.andDo(print())
-                .andExpect(status().isBadRequest())
-                .andExpect(jsonPath("content[0].objectName").exists())
-                .andExpect(jsonPath("content[0].code").exists())
-                .andExpect(jsonPath("content[0].defaultMessage").exists())
-                .andExpect(jsonPath("_links.index").exists())
-        ;
-    }
-    
-    @Test
-    @TestDescription("요청 정보에 해당하는 Event 객체 조회 실패 시, 404 Not Found 처리")
-    @DisplayName("Update Event API : 존재 하지 않는 이벤트 수정 요청")
-    public void updateEventAPI_NotFound_Test() throws Exception {
-        // Given
-        Event event =  this.generatedEvent(100);
-        EventDto eventDto = this.modelMapper.map(event, EventDto.class);
-        String updatedEventName = "updated Event Name";
-        eventDto.setName(updatedEventName);
-
-        // When
-        String urlTemplate = "/api/events/123124";
-        ResultActions resultActions = mockMvc.perform(put(urlTemplate)
-                .header(HttpHeaders.AUTHORIZATION, getBearerToken())
-                .accept(MediaTypes.HAL_JSON_VALUE)
-                .contentType(MediaType.APPLICATION_JSON)
-                .characterEncoding(StandardCharsets.UTF_8.name())
-                .content(this.objectMapper.writeValueAsString(eventDto))
-        );
-
-        // Then
-        resultActions.andDo(print())
-                .andExpect(status().isNotFound())
-                .andExpect(header().exists(HttpHeaders.LOCATION))
-        ;
-    }
-
-    private Event generatedEvent(int index) {
-        Event event = Event.builder()
-                .name("루나소프트 생활 체육회 : " + index)
-                .description("제 2회 루나 배 풋살 대회 : " + index)
-                .beginEnrollmentDateTime(LocalDateTime.of(2020, 8, 6, 9, 30 ))
-                .closeEnrollmentDateTime(LocalDateTime.of(2020, 8, 7, 9, 30 ))
-                .beginEventDateTime(LocalDateTime.of(2020, 8, 13, 19, 0))
-                .endEventDateTime(LocalDateTime.of(2020, 8, 13, 22, 0))
-                .basePrice(100)
-                .maxPrice(200)
-                .limitOfEnrollment(0)
-                .location("서울시 강남구 일원동 마루공원 풋살장 1면")
-                .free(false)
-                .offline(true)
-                .eventStatus(EventStatus.DRAFT)
-                .build();
-
-        Event createdEvent = eventRepository.save(event);
-        return createdEvent;
-    }
-
-
-    private String getBearerToken() throws Exception {
-        return "Bearer " + this.getAccessToken();
-    }
-
-    private String getAccessToken() throws Exception {
-        // Given
-        String clientId = applicationProperties.getCliendId();
-        String clientSecret = applicationProperties.getClinetSecret();
-        String userEmail = applicationProperties.getUserUserName();
-        String userPassword = applicationProperties.getUserPassword();
-        String grantType = applicationProperties.getGrantType();
-
-        Account userAccount = Account.builder()
-                .email(userEmail)
-                .password(userPassword)
-                .roles(Set.of(AccountRole.USER))
-                .build();
-        this.accountService.saveAccount(userAccount);
-
-        // When
-        String urlTemplate = "/oauth/token";
-        ResultActions resultActions = this.mockMvc.perform(post(urlTemplate)
-                .with(httpBasic(clientId, clientSecret)) // clientId와 clientSecret를 이용한 basicOath Header 생성
-                .param("username", userEmail)
-                .param("password", userPassword)
-                .param("grant_type", grantType)
-                .characterEncoding(StandardCharsets.UTF_8.name())
-                .accept(MediaTypes.HAL_JSON_VALUE)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-        );
-
-        String responseBody = resultActions.andReturn().getResponse().getContentAsString();
-        Jackson2JsonParser jackson2JsonParser = new Jackson2JsonParser();
-        return jackson2JsonParser.parseMap(responseBody).get("access_token").toString();
-    }
 }
